Add integration test for simple assumes expression
This is not the ideal test because it depends on that the upstream
charm having the simple

assumes:
  -juju

expression. However, simulating the bug in the facade.py for parsing
such expressions is non-trivial as we need a test to call something
like the CharmsFacade.CharmInfo() to trigger parsing the metadata
(which is actually where it fails in the reported bug). Creating a
local charm wouldn't work because we locally handle the metadata
(wihtout going through anything in the facade.py where the bug is
located). Maybe we can manually call AddCharm in the test for a local
charm and then manually call the CharmInfo with the returned url.