Add example local charm to test and reproduce #865
Unfortunately we can't write an integration test that uses this
because we run our tests on lxd so the charm will be deployed but will
never actually get 'active'. We could test the deployment itself (that
it doesn't error), but if we did that, the 3.0 will actually fail to
deploy on lxd because the series is kubernetes so the test will be
invalid anyways.