Update async tests to use builtin python suite
As of Python 3.8, `AsyncMock` and many other async testing tools are
available as standard libraries. This PR removes the third-party library
`asynctest` and replaces it with those available in the standard python
library.

As part of this, the tests will now run on Python 3.11, which they
previously did not.

Additionally, warnings during tests are significantly reduced (from
hundreds, to tens).