Pass correct charm url to series selector
Fixes #925

Currently we seem to be passing to the series selector the default url
for the charm (e.g. `ch:zinc-k8s`), instead of the url returned from
the ResolveCharms facade call (e.g. `ch:amd64/jammy/zinc-k8s-117`). As
a result we're sending the wrong url (e.g., in #925
`ch:amd64/focal/zinc-k8s-117`) to the `AddPendingResources`, which
results in the wrong image to be selected for the resources for the
charm (e.g., in #925 the incorrect oci-image for the zinc-k8s charm).