Allow switch kwarg in refresh to switch to local charms
Check if the switch param is pointing to a local charm. If it is, run
local_refresh as if it were provided

This is how the Juju CLI works, which was missing from pylibjuju.
Replicate this.

Also push url parsing of switch kwarg behind a check if it's local
charm, meaning we avoid the possibility where we attempt to parse a path
as if it were a url

Resolves https://github.com/juju/python-libjuju/pull/967