Ensure valid charm origin for local charm switches
Ensure that we provide the server with a valid charm origin when
switching to local charms. Previosuly, we simply sent back the same
charm origin when switching to a local charm.

However, this is not valid, since we don't know where we switched from.
If it's a charmhub charm, it's origin will not be valid.

Also, parse the revision from the new charm url the server sends us

This is equivalent to: https://github.com/juju/juju/pull/16434