Use only client version to check the server connection
Currently we spam debug output for any use of pylibjuju where the
"target" != juju controller version. It should be turned into a
ceiling, as it’s pretty normal to use the latest pylibjuju (currently
3.3.0.0) against any 3.x controller (e.g. 3.1). The only time we emit
a warning should be when we think the client should be upgraded to a
more recent version (e.g. running pylibjuju 3.3.0.0 against juju
controller 3.6).