Move the version info into the source code instead of static VERSION file
This fixes #1025 by removing the dependency to the external static
VERSION file from within the project, allowing it to work in
environments where pylibjuju is installed as a dependency library.

Note that this changes the release process. In particular where we
need to manually change the version information is moved into
version.py (the release process document will be updated) This also
removes the VERSION file.