Add retry for the machine ssh
There's a bit of a gap between the time that the machine is assigned
an IP and the ssh service is up and listening, which creates a race
for the ssh command (i.e. we may run the ssh command after the machine
gets the IP but before the ssh service is up). So we retry a couple of
times to mitigate that effect until either the ssh command succeeds,
or we run out of attempts (i.e. something else is going on).