fix(series): add noble support
The pylibjuju client hardcodes a mapping between series codename
and the base channel.

Noble was absent from this mapping, meaning pylibjuju doesn't know
know to deploy to noble.

Add noble to this mapping.

After noble, we will only support deploying by base. Noble will be the
last series added to this map

Add an integration test which deploys a charm to noble