fix(refresh): bug with revisions
We are incorrectly refreshing charms. The current method queries the
controller for the charm's current origin and charm url

In pylibjuju, we then just overwrite the appropriate fields. However,
this leaves behind id_ and hash_

This means when we later refresh, we are always returned the most recent
revision by charmhub, no matter what revision we specify.

The way we handle this in the Juju CLI is by reconstructing the charm
origin using MakeOrigin:
https://github.com/juju/juju/blob/a03ade4d358b1b0f135374b15b1bcd6b46d0ba0f/cmd/juju/application/utils/origin.go#L20-L75

This ensures only the fields we expect are present.

Replicate this method, ensuring we behave the same as the CLI.

Fixes: https://github.com/juju/python-libjuju/issues/1057